refactor(audiobook): extract stateless PlayerSurface for reuse
